+++ Changelog.md
@@ -11,3 +11,4 @@ Bugfixes:
 * Button color standardization in song view mode switch (#99)
 * Standardized and extended additional information that can be shown on slides: Song Books, Author, Publisher, Copyright and CCLI ID (see Program Settings dialog -> Master formatting). (#100, #101)
 * Double-click on song in song statistics dialog opens song in editor (#103)
+* Window and columns of metadata editor can now be resized (#104)